like functionality and auth system completed

+++ server/src/handlers/posts.js
@@ -1,9 +1,19 @@
+import mongoose from "mongoose";
 import PostModel from "../models/posts.js";
 
 const addPosts = async (req, res) => {
-  const body = req.body;
-  const post = new PostModel(body);
   try {
+    const body = req.body;
+
+    const post = new PostModel({
+      ...body,
+      creator: req?.userId,
+      author: req?.userName,
+      createdAt: new Date().toISOString(),
+    });
+
+    if (!req.userId) return res.json({ message: "Unauthenticated !" });
+
     const result = await post.save();
     res.status(201).json(result);
   } catch (error) {
@@ -15,7 +25,7 @@ const addPosts = async (req, res) => {
 
 const getPosts = async (req, res) => {
   try {
-    const result = await PostModel.find();
+    const result = await PostModel.find().sort({ createdAt: -1 });
     res.status(201).json(result);
   } catch (error) {
     res.status(401).json({
@@ -25,8 +35,14 @@ const getPosts = async (req, res) => {
 };
 
 const updatePosts = async (req, res) => {
-  const id = req.params.id;
   try {
+    const id = req.params.id;
+
+    if (!req.userId) return res.json({ message: "Unauthenticated !" });
+
+    if (!mongoose.Types.ObjectId.isValid(id))
+      return res.status(404).send("No Post with that id");
+
     const updatedPost = await PostModel.findByIdAndUpdate(
       { _id: id },
       req.body,
@@ -43,8 +59,14 @@ const updatePosts = async (req, res) => {
 };
 
 const deletePosts = async (req, res) => {
-  const id = req.params.id;
   try {
+    const id = req.params.id;
+
+    if (!req.userId) return res.json({ message: "Unauthenticated !" });
+
+    if (!mongoose.Types.ObjectId.isValid(id))
+      return res.status(404).send("No Post with that id");
+
     const result = await PostModel.findByIdAndRemove({ _id: id }, req.body);
     res.status(201).json(id);
   } catch (error) {
@@ -54,4 +76,32 @@ const deletePosts = async (req, res) => {
   }
 };
 
-export { addPosts, getPosts, updatePosts, deletePosts };
+const likePosts = async (req, res) => {
+  const id = req.params.id;
+  try {
+    if (!req.userId) return res.json({ message: "Unauthenticated !" });
+
+    if (!mongoose.Types.ObjectId.isValid(id))
+      return res.status(404).send("No Post with that id");
+
+    const thePost = await PostModel.findById(id);
+    const alreadyLiked = thePost.likes.indexOf(req.userId);
+
+    if (alreadyLiked === -1) {
+      thePost.likes.push(req.userId);
+    } else {
+      thePost.likes = thePost.likes.filter((id) => id !== req.userId);
+    }
+
+    const likesUpdatedPost = await PostModel.findByIdAndUpdate(id, thePost, {
+      new: true,
+    });
+    res.status(201).json(likesUpdatedPost);
+  } catch (error) {
+    res.status(401).json({
+      message: error.message,
+    });
+  }
+};
+
+export { addPosts, getPosts, updatePosts, deletePosts, likePosts };
